added little style

+++ src/ui/InvoiceItem.tsx
@@ -70,8 +70,23 @@ function InvoiceItem() {
     });
   }, [queryClient]);
 
+
+
+  const pageStyle = `@page {
+    size: A4 landscape;
+    // background : var(--bg-color-ter);
+    // backgroundColor : #1d2238;
+  }`;
+
+
+
+
   const handlePrint = useReactToPrint({
     content: () => componentRef.current,
+    documentTitle: "Invoice",
+    pageStyle() {
+      return pageStyle;
+    },
   });
 
   if (isLoading && userLoading)
@@ -89,8 +104,10 @@ function InvoiceItem() {
     navigate(`/invoice/edit/${id}`);
   }
   function deleteInvoice() {
-    //@ts-ignore
-    deleteMutate(id);
+    if (confirm("Do you want to delete this Invoice")) {
+      //@ts-ignore
+      deleteMutate(id);
+    }
   }
 
   function markAsRead() {
@@ -99,8 +116,9 @@ function InvoiceItem() {
   }
 
   return (
-    <section className=" bg-[var(--bg-color-ter)] w-full h-[100vh] pt-10  text-white pb-4 overflow-scroll flow">
-      <div className="max-w-[35rem] w-[90%] m-auto">
+    // I changed height here
+    <section className=" bg-[var(--bg-color-ter)] w-full h-[100%] pt-10  text-white pb-4 overflow-scroll flow">
+      <div className="max-w-[55rem] w-[95%] m-auto">
         <p
           className="flex items-center gap-1 mb-4 cursor-pointer text-[var(--color-text-white)]"
           onClick={() => navigate("/")}
@@ -108,67 +126,73 @@ function InvoiceItem() {
           <FaArrowLeftLong />
           Back
         </p>
-        <div ref={componentRef} className={`bg-[var(--bg-color-ter)] p-2`}>
-          <article
-            className={`bg-[var(--bg-color-primary)] ${
-              isDark && "shadow-xl"
-            } rounded-md p-4 mb-8`}
+          <div
+            ref={componentRef}
+            className={`bg-[var(--bg-color-ter)] p-2`}
           >
-            <div className="flex justify-between items-center">
-              <div className="flex items-center gap-x-4 text-sm">
-                <span className="text-lg text-[var(--color-text-white)]">
-                  Status
-                </span>
-                {/* @ts-ignore */}
-                <LabelContainer type={`${invoice.status}`}>
-                  {/* @ts-ignore */}
-                  <Labels type={`${invoice.status}`}>{invoice.status}</Labels>
-                </LabelContainer>
-              </div>
-              {currentUser?.isAdmin ? (
-                <div className="flex">
-                  {invoice.status === "paid" ? (
-                    ""
-                  ) : (
-                    <Button type="edit" onClick={editInvoice}>
-                      Edit
-                    </Button>
-                  )}
-                  <Button type="delete" onClick={deleteInvoice}>
-                    Delete
-                  </Button>
-                  {invoice.status === "paid" ? (
-                    ""
-                  ) : (
-                    <Button type="mark" onClick={markAsRead}>
-                      Mark as Paid
-                    </Button>
-                  )}
-                </div>
-              ) : currentUser?._id === invoice.posterId ? (
-                <>
-                  {invoice.status === "paid" ? (
-                    ""
-                  ) : (
-                    <Button type="edit" onClick={editInvoice}>
-                      Edit
-                    </Button>
-                  )}
-                  <Button
-                    type="delete"
-                    disabled={isDeleting}
-                    onClick={deleteInvoice}
+            <article
+              className={`bg-[var(--bg-color-primary)] ${
+                isDark && "shadow-xl"
+              } rounded-md p-4 mb-8`}
+            >
+              <div className="flex sm:justify-between justify-center items-center flex-wrap gap-4">
+                <div className="flex items-center gap-x-4 text-sm">
+                  <span
+                    className="md:text-lg text-base text-[var(--color-text-white)]"
                   >
-                    Delete
-                  </Button>
-                </>
-              ) : (
-                ""
-              )}
-            </div>
-          </article>
-          <Invoice invoice={invoice} />
-        </div>
+                    Status
+                  </span>
+                  {/* @ts-ignore */}
+                  <LabelContainer type={`${invoice.status}`}>
+                    {/* @ts-ignore */}
+                    <Labels type={`${invoice.status}`}>{invoice.status}</Labels>
+                  </LabelContainer>
+                </div>
+                {currentUser?.isAdmin ? (
+                  <div className="flex justify-around md:justify-normal">
+                    {invoice.status === "paid" ? (
+                      ""
+                    ) : (
+                      <Button type="edit" onClick={editInvoice}>
+                        Edit
+                      </Button>
+                    )}
+                    <Button type="delete" onClick={deleteInvoice}>
+                      Delete
+                    </Button>
+                    {invoice.status === "paid" ? (
+                      ""
+                    ) : (
+                      <Button type="mark" onClick={markAsRead}>
+                        Mark as Paid
+                      </Button>
+                    )}
+                  </div>
+                ) : currentUser?._id === invoice.posterId ? (
+                  <>
+                    {invoice.status === "paid" ? (
+                      ""
+                    ) : (
+                      <Button type="edit" onClick={editInvoice}>
+                        Edit
+                      </Button>
+                    )}
+                    <Button
+                      type="delete"
+                      disabled={isDeleting}
+                      onClick={deleteInvoice}
+                    >
+                      Delete
+                    </Button>
+                  </>
+                ) : (
+                  ""
+                )}
+              </div>
+            </article>
+            <Invoice invoice={invoice} />
+          </div>
+
         <div className="flex justify-center">
           <button className="p-2 bg-[#7c5df9] my-4" onClick={handlePrint}>
             Print
